Show root mode pref on non-rooted devices
Continuing the theme of "don't prompt people for root when not needed",
this time on the preferences screen — we only want to prompt if they
actually enable root mode, not when they open the preferences screen.

In order to do this, we need to make the root mode preference visible to
people on non-rooted phones. This is a good thing! Otherwise those
people wouldn't know the feature exists; maybe some of them will be
motivated to root their phone!