Only send preview toggle intent when needed
We had an issue that kept popping up: When you slid the preview bar and
the service couldn't start for some reason, it would start the service
over and over and over again. This would result in, for example, a
torrent of root mode permission prompts.

It happened because we sent the intent each time a touch was detected on
the slidebar.  We don't need to do that, only when the first touch is
registered and the last touch is released.